Implement username conversion and validation schema using Zod in multiple components
